+++ config.py
@@ -10,3 +10,9 @@ SQLALCHEMY_MIGRATE_REPO = os.path.join(basedir, 'db_repository')
 DEBUG = False
 SQLALCHEMY_ECHO = False
 SQLALCHEMY_TRACK_MODIFICATIONS = True
+
+# Constant to set interval of requests:
+REQUEST_INTERVAL = 10
+
+# Time in seconds while data is valuable:
+EXPIRATION_INTERVAL = 1800
