+++ create_db.py
@@ -1,9 +1,12 @@
 #!flask/bin/python
+import os.path
+
 from migrate.versioning import api
+
+from app import db
 from config import SQLALCHEMY_DATABASE_URI
 from config import SQLALCHEMY_MIGRATE_REPO
-from app import db
-import os.path
+
 db.create_all()
 if not os.path.exists(SQLALCHEMY_MIGRATE_REPO):
     api.create(SQLALCHEMY_MIGRATE_REPO, 'database repository')
